refactor(e2e): merge duplicate-slash tests and apply slashing presets
Merge duplicate_attestation_slash + duplicate_proposal_slash into
slashing/equivocation_slash.parallel.test.ts (shared harness opts, per-it setup),
preserving the proposal test's all-node offense poll + proposer-for-slot check verbatim.
Apply withOnlyOffense('slashDuplicateProposalPenalty') to equivocation (replacing its
~9-line manual penalty zero-out) and MOCK_GOSSIP_MULTI_VALIDATOR_OPTS +
buildMockGossipValidators to equivocation and invalidate_block (penalties left explicit
where the test relies on config defaults). Update .test_patterns.yml flake entry to the
merged file. The remaining ~10 e2e_p2p slashing/sentinel conversions are deferred: they
run real libp2p (P2PNetworkTest/createNodes) and the heaviest ones stub libp2p internals
the mock bus cannot reproduce — genuine work, not a folder move.
